zstd: x86 assembler implementation of sequenceDecs.decode
Differences with the Go implementation:
- check ml and mo in the main loop,
- s.seqSize and litRemain are checked in the end.
