+++ zstd/autogen.go
@@ -1,0 +1,4 @@
+package zstd
+
+//go:generate go run generate.go
+//go:generate asmfmt -w seqdec_amd64.s
